4o: day 15 part 1 - succeeds with reprompt
### Reprompt

Careful in #selection not to move a box if it would intersect with a wall. Also, any number of boxes in a row can be moved, so long as the end one isn't next to a wall.
(For #selection, I just picked the condition and body of the if statement on line 50